+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## _Unreleased_
+
+## v0.3.0
+
 - Prevent HK iframes eating damage dealt in quick succession
 - Fix bind/taunt being broken after dialogue
 - Fix not being able to open inventory on controller